requirements.txt
asgiref==3.4.1
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
click==8.1.3
decorator==5.1.1
Django==3.2.16
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-haystack==3.1.1
djangorestframework==3.12.4
drf-haystack==1.8.11
elasticsearch==7.17.6
executing==1.1.0
factory-boy==3.2.1
Faker==14.2.1
flake8==5.0.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
typing-extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
